Struct rustc_index::bit_set::BitMatrix

pub struct BitMatrix<R: Idx, C: Idx> {
    num_rows: usize,
    num_columns: usize,
    words: SmallVec<[u64; 2]>,
    marker: PhantomData<(R, C)>,
}

A fixed-size 2D bit matrix type with a dense representation.

R and C are index types used to identify rows and columns respectively; typically newtyped usize wrappers, but they can also just be usize.

All operations that involve a row and/or column index will panic if the index exceeds the relevant bound.

Fields
num_rows: usizenum_columns: usizewords: SmallVec<[u64; 2]>marker: PhantomData<(R, C)>
Implementations

impl<R: Idx, C: Idx> BitMatrix<R, C>

pub fn new(num_rows: usize, num_columns: usize) -> BitMatrix<R, C>

Creates a new rows x columns matrix, initially empty.

pub fn from_row_n(row: &BitSet<C>, num_rows: usize) -> BitMatrix<R, C>

Creates a new matrix, with row used as the value for every row.

pub fn rows(&self) -> impl Iterator<Item = R>

fn range(&self, row: R) -> (usize, usize)

The range of bits for a given row.

pub fn insert(&mut self, row: R, column: C) -> bool

Sets the cell at (row, column) to true. Put another way, insert column to the bitset for row.

Returns true if this changed the matrix.

pub fn contains(&self, row: R, column: C) -> bool

Do the bits from row contain column? Put another way, is the matrix cell at (row, column) true? Put yet another way, if the matrix represents (transitive) reachability, can row reach column?

pub fn intersect_rows(&self, row1: R, row2: R) -> Vec<C>

Returns those indices that are true in rows a and b. This is an O(n) operation where n is the number of elements (somewhat independent from the actual size of the intersection, in particular).

pub fn union_rows(&mut self, read: R, write: R) -> bool

Adds the bits from row read to the bits from row write, and returns true if anything changed.

This is used when computing transitive reachability because if you have an edge write -> read, because in that case write can reach everything that read can (and potentially more).

pub fn union_row_with(&mut self, with: &BitSet<C>, write: R) -> bool

Adds the bits from with to the bits from row write, and returns true if anything changed.

pub fn insert_all_into_row(&mut self, row: R)

Sets every cell in row to true.

pub fn words(&self) -> &[u64]

Gets a slice of the underlying words.

pub fn iter(&self, row: R) -> BitIter<'_, C>

Iterates through all the columns set to true in a given row of the matrix.

pub fn count(&self, row: R) -> usize

Returns the number of elements in row.