Struct rustc_index::bit_set::BitSet

pub struct BitSet<T> {
    domain_size: usize,
    words: Vec<u64>,
    marker: PhantomData<T>,
}

A fixed-size bitset type with a dense representation.

NOTE: Use GrowableBitSet if you need support for resizing after creation.

T is an index type, typically a newtyped usize wrapper, but it can also just be usize.

All operations that involve an element will panic if the element is equal to or greater than the domain size. All operations that involve two bitsets will panic if the bitsets have differing domain sizes.

Fields
domain_size: usizewords: Vec<u64>marker: PhantomData<T>
Implementations

impl<T> BitSet<T>

pub fn domain_size(&self) -> usize

Gets the domain size.

impl<T: Idx> BitSet<T>

pub fn new_empty(domain_size: usize) -> BitSet<T>

Creates a new, empty bitset with a given domain_size.

pub fn new_filled(domain_size: usize) -> BitSet<T>

Creates a new, filled bitset with a given domain_size.

pub fn clear(&mut self)

Clear all elements.

fn clear_excess_bits(&mut self)

Clear excess bits in the final word.

pub fn count(&self) -> usize

Count the number of set bits in the set.

pub fn contains(&self, elem: T) -> bool

Returns true if self contains elem.

pub fn superset(&self, other: &BitSet<T>) -> bool

Is self is a (non-strict) superset of other?

pub fn is_empty(&self) -> bool

Is the set empty?

pub fn insert(&mut self, elem: T) -> bool

Insert elem. Returns whether the set has changed.

pub fn insert_range(&mut self, elems: impl RangeBounds<T>)

pub fn insert_all(&mut self)

Sets all bits to true.

pub fn remove(&mut self, elem: T) -> bool

Returns true if the set has changed.

pub fn words(&self) -> &[u64]

Gets a slice of the underlying words.

pub fn iter(&self) -> BitIter<'_, T>

Iterates over the indices of set bits in a sorted order.

pub fn to_hybrid(&self) -> HybridBitSet<T>

Duplicates the set as a hybrid set.

fn reverse_union_sparse(&mut self, sparse: &SparseBitSet<T>) -> bool

Set self = self | other. In contrast to union returns true if the set contains at least one bit that is not in other (i.e. other is not a superset of self).

This is an optimization for union of a hybrid bitset.

fn last_set_in(&self, range: impl RangeBounds<T>) -> Option<T>

pub fn union<Rhs>(&mut self, other: &Rhs) -> boolwhere
Self: BitRelations<Rhs>,

Sets self = self | other and returns true if self changed (i.e., if new bits were added).

pub fn subtract<Rhs>(&mut self, other: &Rhs) -> boolwhere
Self: BitRelations<Rhs>,

Sets self = self - other and returns true if self changed. (i.e., if any bits were removed).

pub fn intersect<Rhs>(&mut self, other: &Rhs) -> boolwhere
Self: BitRelations<Rhs>,

Sets self = self & other and return true if self changed. (i.e., if any bits were removed).