Struct rustc_index::bit_set::ChunkedBitSet

pub struct ChunkedBitSet<T> {
    domain_size: usize,
    chunks: Box<[Chunk]>,
    marker: PhantomData<T>,
}

A fixed-size bitset type with a partially dense, partially sparse representation. The bitset is broken into chunks, and chunks that are all zeros or all ones are represented and handled very efficiently.

This type is especially efficient for sets that typically have a large domain_size with significant stretches of all zeros or all ones, and also some stretches with lots of 0s and 1s mixed in a way that causes trouble for IntervalSet.

T is an index type, typically a newtyped usize wrapper, but it can also just be usize.

All operations that involve an element will panic if the element is equal to or greater than the domain size. All operations that involve two bitsets will panic if the bitsets have differing domain sizes.

Fields
domain_size: usizechunks: Box<[Chunk]>

The chunks. Each one contains exactly CHUNK_BITS values, except the last one which contains 1..=CHUNK_BITS values.

marker: PhantomData<T>
Implementations

impl<T> ChunkedBitSet<T>

pub fn domain_size(&self) -> usize

impl<T: Idx> ChunkedBitSet<T>

fn new(domain_size: usize, is_empty: bool) -> Self

Creates a new bitset with a given domain_size and chunk kind.

pub fn new_empty(domain_size: usize) -> Self

Creates a new, empty bitset with a given domain_size.

pub fn new_filled(domain_size: usize) -> Self

Creates a new, filled bitset with a given domain_size.

pub fn count(&self) -> usize

Count the number of bits in the set.

pub fn contains(&self, elem: T) -> bool

Returns true if self contains elem.

pub fn iter(&self) -> ChunkedBitIter<'_, T>

pub fn insert(&mut self, elem: T) -> bool

Insert elem. Returns whether the set has changed.

pub fn insert_all(&mut self)

Sets all bits to true.

pub fn remove(&mut self, elem: T) -> bool

Returns true if the set has changed.

pub fn union<Rhs>(&mut self, other: &Rhs) -> boolwhere
Self: BitRelations<Rhs>,

Sets self = self | other and returns true if self changed (i.e., if new bits were added).

pub fn subtract<Rhs>(&mut self, other: &Rhs) -> boolwhere
Self: BitRelations<Rhs>,

Sets self = self - other and returns true if self changed. (i.e., if any bits were removed).

pub fn intersect<Rhs>(&mut self, other: &Rhs) -> boolwhere
Self: BitRelations<Rhs>,

Sets self = self & other and return true if self changed. (i.e., if any bits were removed).