Struct rustc_index::vec::IndexVec

pub struct IndexVec<I: Idx, T> {
    pub raw: Vec<T>,
    _marker: PhantomData<fn(_: &I)>,
}

Fields
raw: Vec<T>_marker: PhantomData<fn(_: &I)>
Implementations

impl<I: Idx, T> IndexVec<I, T>

pub fn new() -> Self

pub fn from_raw(raw: Vec<T>) -> Self

pub fn with_capacity(capacity: usize) -> Self

pub fn from_elem<S>(elem: T, universe: &IndexVec<I, S>) -> Selfwhere
T: Clone,

pub fn from_elem_n(elem: T, n: usize) -> Selfwhere
T: Clone,

pub fn from_fn_n(func: impl FnMut(I) -> T, n: usize) -> Self

Create an IndexVec with n elements, where the value of each element is the result of func(i). (The underlying vector will be allocated only once, with a capacity of at least n.)

pub fn push(&mut self, d: T) -> I

pub fn pop(&mut self) -> Option<T>

pub fn len(&self) -> usize

pub fn next_index(&self) -> I

Gives the next index that will be assigned when push is called.

pub fn is_empty(&self) -> bool

pub fn into_iter(self) -> IntoIter<T>

pub fn into_iter_enumerated(
self
) -> impl DoubleEndedIterator<Item = (I, T)> + ExactSizeIterator

pub fn iter(&self) -> Iter<'_, T>

pub fn iter_enumerated(
&self
) -> impl DoubleEndedIterator<Item = (I, &T)> + ExactSizeIterator + '_

pub fn indices(
&self
) -> impl DoubleEndedIterator<Item = I> + ExactSizeIterator + Clone + 'static

pub fn iter_mut(&mut self) -> IterMut<'_, T>

pub fn iter_enumerated_mut(
&mut self
) -> impl DoubleEndedIterator<Item = (I, &mut T)> + ExactSizeIterator + '_

pub fn drain<'a, R: RangeBounds<usize>>(
&'a mut self,
range: R
) -> impl Iterator<Item = T> + 'a

pub fn drain_enumerated<'a, R: RangeBounds<usize>>(
&'a mut self,
range: R
) -> impl Iterator<Item = (I, T)> + 'a

pub fn last(&self) -> Option<I>

pub fn shrink_to_fit(&mut self)

pub fn swap(&mut self, a: I, b: I)

pub fn truncate(&mut self, a: usize)

pub fn get(&self, index: I) -> Option<&T>

pub fn get_mut(&mut self, index: I) -> Option<&mut T>

pub fn pick2_mut(&mut self, a: I, b: I) -> (&mut T, &mut T)

Returns mutable references to two distinct elements, a and b.

Panics if a == b.

pub fn pick3_mut(&mut self, a: I, b: I, c: I) -> (&mut T, &mut T, &mut T)

Returns mutable references to three distinct elements.

Panics if the elements are not distinct.

pub fn convert_index_type<Ix: Idx>(self) -> IndexVec<Ix, T>

pub fn ensure_contains_elem(&mut self, elem: I, fill_value: impl FnMut() -> T)

Grows the index vector so that it contains an entry for elem; if that is already true, then has no effect. Otherwise, inserts new values as needed by invoking fill_value.

pub fn resize_to_elem(&mut self, elem: I, fill_value: impl FnMut() -> T)

impl<I: Idx, T> IndexVec<I, Option<T>>

IndexVec is often used as a map, so it provides some map-like APIs.

pub fn insert(&mut self, index: I, value: T) -> Option<T>

pub fn get_or_insert_with(
&mut self,
index: I,
value: impl FnOnce() -> T
) -> &mut T

pub fn remove(&mut self, index: I) -> Option<T>

impl<I: Idx, T: Clone> IndexVec<I, T>

pub fn resize(&mut self, new_len: usize, value: T)

impl<I: Idx, T: Ord> IndexVec<I, T>

pub fn binary_search(&self, value: &T) -> Result<I, I>