Crate rustc_index

Re-exports

pub use vec::IndexVec;

Modules

bit_set

idx 🔒

interval

slice 🔒

vec 🔒

Macros

newtype_index

Creates a struct type S that can be used as an index with IndexVec and so on.

static_assert_size

Type size assertion. The first argument is a type and the second argument is its expected size.

Structs

IndexSlice

A view into contiguous Ts, indexed by I rather than by usize.

Traits

Idx

Represents some newtyped usize wrapper.